pyproject.toml
[tool.poetry]
name = "saber-test"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
fastapi = "^0.104.1"
uvicorn = "^0.24.0.post1"
pyyaml = "^6.0.1"
pydantic = "^2.5.1"
pytest = "^7.4.3"
httpx = "^0.25.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
